refactor: harden HTML import/export feature
- Add span to INLINE_TEXT_TAGS to prevent nested <p> in rendered output
- Extract shared escapeHtml, tailwind-colors, and tailwind-normalizer
- Fix br incorrectly mapped to hr; skip br at block level
- Add textarea/select/form attribute capture on import and export
- Add overflow class parsing to classesToDesign
- Add background-image, flex-grow/shrink/basis, order to styleToClasses
- Sanitize SVG on import by stripping script tags and on* attributes
- Wrap htmlToLayers and handleImportHtml in try/catch error handling
- Replace deprecated execCommand copy fallback with toast message
- Widen makeTextLayer to accept string|object, remove inline duplicate

Made-with: Cursor
