+++ lib/escape-html.ts
@@ -1,0 +1,12 @@
+/**
+ * Escape HTML special characters to prevent XSS.
+ * Shared utility used by page-fetcher, html-layer-converter, and emailService.
+ */
+export function escapeHtml(str: string): string {
+  return str
+    .replace(/&/g, '&amp;')
+    .replace(/</g, '&lt;')
+    .replace(/>/g, '&gt;')
+    .replace(/"/g, '&quot;')
+    .replace(/'/g, '&#039;');
+}
