+++ lib/html-layer-converter.ts
@@ -12,6 +12,8 @@ import { generateId } from '@/lib/utils';
 import { classesToDesign } from '@/lib/tailwind-class-mapper';
 import { getClassesString, getLayerHtmlTag } from '@/lib/layer-utils';
 import { getTiptapTextContent } from '@/lib/text-format-utils';
+import { escapeHtml } from '@/lib/escape-html';
+import { normalizeV3ToV4, resolveNamedColors } from '@/lib/tailwind-normalizer';
 
 // ─── Tag → Layer Name Mapping ───
 
@@ -132,7 +134,6 @@ const TAG_TO_LAYER_NAME: Record<string, string> = {
 
   // Separators
   hr: 'hr',
-  br: 'hr',
 
   // Preformatted
   pre: 'div',
@@ -186,209 +187,8 @@ const INLINE_TEXT_TAGS = new Set([
   'strong', 'b', 'em', 'i', 'u', 's', 'del', 'ins', 'mark', 'small',
   'sub', 'sup', 'abbr', 'cite', 'code', 'kbd', 'samp', 'var', 'time',
   'data', 'q', 'dfn', 'ruby', 'rt', 'rp', 'bdi', 'bdo', 'wbr',
-  'a',
+  'a', 'span',
 ]);
-
-// ─── Tailwind v3 → v4 Normalizer ───
-
-const V3_TO_V4_RENAMES: Record<string, string> = {
-  'flex-grow': 'grow',
-  'flex-grow-0': 'grow-0',
-  'flex-shrink': 'shrink',
-  'flex-shrink-0': 'shrink-0',
-  'overflow-ellipsis': 'text-ellipsis',
-  'decoration-clone': 'box-decoration-clone',
-  'decoration-slice': 'box-decoration-slice',
-  'outline-none': 'outline-hidden',
-  'ring': 'ring-3',
-  'shadow-inner': 'inset-shadow-sm',
-  'break-words': 'wrap-break-word',
-  'transform': '',
-  'filter': '',
-  'backdrop-filter': '',
-};
-
-const OPACITY_RE = /^(bg|text|border|ring|divide|placeholder)-opacity-(?:(\d+)|\[(\d+)%?\])$/;
-const NAMED_COLOR_RE = /^[a-z]+-\d{2,3}$/;
-const SIZE_VALUE_RE = /^\d|(?:px|em|rem|%|vh|vw|ch|ex|svh|dvh|lvh)$/;
-
-const VARIANT_PREFIX_RE = /^((?:[a-z0-9-]+:)+)/;
-
-function splitVariantPrefix(cls: string): [string, string] {
-  const m = cls.match(VARIANT_PREFIX_RE);
-  return m ? [m[1], cls.slice(m[1].length)] : ['', cls];
-}
-
-function isArbitraryColorClass(_prefix: string, rest: string): boolean {
-  if (!rest.startsWith('[')) return false;
-  const inner = rest.slice(1, rest.indexOf(']'));
-  if (SIZE_VALUE_RE.test(inner)) return false;
-  return true;
-}
-
-const V3_TO_V4_SCALE: Record<string, string> = {
-  '': 'sm', 'sm': 'xs',
-};
-
-const ROUNDED_RE = /^rounded(-(?:t|r|b|l|tl|tr|bl|br|s|e|ss|se|es|ee))?(?:-(sm|md|lg|xl|2xl|3xl))?$/;
-
-function renameScaledUtility(cls: string): string | null {
-  const roundedMatch = cls.match(ROUNDED_RE);
-  if (roundedMatch) {
-    const dir = roundedMatch[1] || '';
-    const size = roundedMatch[2] || '';
-    if (!(size in V3_TO_V4_SCALE)) return null;
-    const v4 = V3_TO_V4_SCALE[size];
-    return `rounded${dir}${v4 ? `-${v4}` : ''}`;
-  }
-
-  for (const pfx of ['drop-shadow', 'shadow', 'backdrop-blur', 'blur'] as const) {
-    if (cls === pfx || cls.startsWith(`${pfx}-`)) {
-      const rest = cls.slice(pfx.length);
-      const m = rest.match(/^(?:-(sm|md|lg|xl|2xl|3xl))?$/);
-      if (!m) return null;
-      const size = m[1] || '';
-      if (!(size in V3_TO_V4_SCALE)) return null;
-      const v4 = V3_TO_V4_SCALE[size];
-      return `${pfx}${v4 ? `-${v4}` : ''}`;
-    }
-  }
-
-  return null;
-}
-
-function normalizeV3ToV4(classes: string[]): string[] {
-  const result: string[] = [];
-  const opacityEntries: { prefix: string; value: string }[] = [];
-
-  for (const cls of classes) {
-    const [variantPrefix, base] = splitVariantPrefix(cls);
-
-    const renamed = V3_TO_V4_RENAMES[base];
-    if (renamed !== undefined) {
-      if (renamed) result.push(variantPrefix + renamed);
-      continue;
-    }
-
-    const opMatch = base.match(OPACITY_RE);
-    if (opMatch) {
-      const opValue = opMatch[2] || opMatch[3];
-      if (opValue && opValue !== '100') {
-        opacityEntries.push({ prefix: opMatch[1], value: opValue });
-      }
-      continue;
-    }
-
-    if (base.startsWith('bg-gradient-to-')) {
-      result.push(variantPrefix + 'bg-linear-to-' + base.slice(15));
-      continue;
-    }
-
-    const scaled = renameScaledUtility(base);
-    if (scaled) {
-      result.push(variantPrefix + scaled);
-      continue;
-    }
-
-    result.push(cls);
-  }
-
-  for (const { prefix, value } of opacityEntries) {
-    const colorIdx = result.findIndex(cls => {
-      if (cls.includes('/')) return false;
-      if (!cls.startsWith(`${prefix}-`)) return false;
-      const rest = cls.slice(prefix.length + 1);
-      if (rest.startsWith('[')) return isArbitraryColorClass(prefix, rest);
-      return NAMED_COLOR_RE.test(rest);
-    });
-
-    if (colorIdx !== -1) {
-      result[colorIdx] = `${result[colorIdx]}/${value}`;
-    }
-  }
-
-  for (let i = 0; i < result.length; i++) {
-    const [vp, base] = splitVariantPrefix(result[i]);
-    if (base.startsWith('placeholder-') && !base.includes(':')) {
-      result[i] = `${vp}placeholder:text-${base.slice(12)}`;
-    }
-  }
-
-  return result;
-}
-
-// ─── Named Tailwind Color → Hex Resolver ───
-
-const TAILWIND_COLORS: Record<string, Record<string, string>> = {
-  slate: { '50':'#f8fafc','100':'#f1f5f9','200':'#e2e8f0','300':'#cbd5e1','400':'#94a3b8','500':'#64748b','600':'#475569','700':'#334155','800':'#1e293b','900':'#0f172a','950':'#020617' },
-  gray: { '50':'#f9fafb','100':'#f3f4f6','200':'#e5e7eb','300':'#d1d5db','400':'#9ca3af','500':'#6b7280','600':'#4b5563','700':'#374151','800':'#1f2937','900':'#111827','950':'#030712' },
-  zinc: { '50':'#fafafa','100':'#f4f4f5','200':'#e4e4e7','300':'#d4d4d8','400':'#a1a1aa','500':'#71717a','600':'#52525b','700':'#3f3f46','800':'#27272a','900':'#18181b','950':'#09090b' },
-  neutral: { '50':'#fafafa','100':'#f5f5f5','200':'#e5e5e5','300':'#d4d4d4','400':'#a3a3a3','500':'#737373','600':'#525252','700':'#404040','800':'#262626','900':'#171717','950':'#0a0a0a' },
-  stone: { '50':'#fafaf9','100':'#f5f5f4','200':'#e7e5e4','300':'#d6d3d1','400':'#a8a29e','500':'#78716c','600':'#57534e','700':'#44403c','800':'#292524','900':'#1c1917','950':'#0c0a09' },
-  red: { '50':'#fef2f2','100':'#fee2e2','200':'#fecaca','300':'#fca5a5','400':'#f87171','500':'#ef4444','600':'#dc2626','700':'#b91c1c','800':'#991b1b','900':'#7f1d1d','950':'#450a0a' },
-  orange: { '50':'#fff7ed','100':'#ffedd5','200':'#fed7aa','300':'#fdba74','400':'#fb923c','500':'#f97316','600':'#ea580c','700':'#c2410c','800':'#9a3412','900':'#7c2d12','950':'#431407' },
-  amber: { '50':'#fffbeb','100':'#fef3c7','200':'#fde68a','300':'#fcd34d','400':'#fbbf24','500':'#f59e0b','600':'#d97706','700':'#b45309','800':'#92400e','900':'#78350f','950':'#451a03' },
-  yellow: { '50':'#fefce8','100':'#fef9c3','200':'#fef08a','300':'#fde047','400':'#facc15','500':'#eab308','600':'#ca8a04','700':'#a16207','800':'#854d0e','900':'#713f12','950':'#422006' },
-  lime: { '50':'#f7fee7','100':'#ecfccb','200':'#d9f99d','300':'#bef264','400':'#a3e635','500':'#84cc16','600':'#65a30d','700':'#4d7c0f','800':'#3f6212','900':'#365314','950':'#1a2e05' },
-  green: { '50':'#f0fdf4','100':'#dcfce7','200':'#bbf7d0','300':'#86efac','400':'#4ade80','500':'#22c55e','600':'#16a34a','700':'#15803d','800':'#166534','900':'#14532d','950':'#052e16' },
-  emerald: { '50':'#ecfdf5','100':'#d1fae5','200':'#a7f3d0','300':'#6ee7b7','400':'#34d399','500':'#10b981','600':'#059669','700':'#047857','800':'#065f46','900':'#064e3b','950':'#022c22' },
-  teal: { '50':'#f0fdfa','100':'#ccfbf1','200':'#99f6e4','300':'#5eead4','400':'#2dd4bf','500':'#14b8a6','600':'#0d9488','700':'#0f766e','800':'#115e59','900':'#134e4a','950':'#042f2e' },
-  cyan: { '50':'#ecfeff','100':'#cffafe','200':'#a5f3fc','300':'#67e8f9','400':'#22d3ee','500':'#06b6d4','600':'#0891b2','700':'#0e7490','800':'#155e75','900':'#164e63','950':'#083344' },
-  sky: { '50':'#f0f9ff','100':'#e0f2fe','200':'#bae6fd','300':'#7dd3fc','400':'#38bdf8','500':'#0ea5e9','600':'#0284c7','700':'#0369a1','800':'#075985','900':'#0c4a6e','950':'#082f49' },
-  blue: { '50':'#eff6ff','100':'#dbeafe','200':'#bfdbfe','300':'#93c5fd','400':'#60a5fa','500':'#3b82f6','600':'#2563eb','700':'#1d4ed8','800':'#1e40af','900':'#1e3a8a','950':'#172554' },
-  indigo: { '50':'#eef2ff','100':'#e0e7ff','200':'#c7d2fe','300':'#a5b4fc','400':'#818cf8','500':'#6366f1','600':'#4f46e5','700':'#4338ca','800':'#3730a3','900':'#312e81','950':'#1e1b4b' },
-  violet: { '50':'#f5f3ff','100':'#ede9fe','200':'#ddd6fe','300':'#c4b5fd','400':'#a78bfa','500':'#8b5cf6','600':'#7c3aed','700':'#6d28d9','800':'#5b21b6','900':'#4c1d95','950':'#2e1065' },
-  purple: { '50':'#faf5ff','100':'#f3e8ff','200':'#e9d5ff','300':'#d8b4fe','400':'#c084fc','500':'#a855f7','600':'#9333ea','700':'#7e22ce','800':'#6b21a8','900':'#581c87','950':'#3b0764' },
-  fuchsia: { '50':'#fdf4ff','100':'#fae8ff','200':'#f5d0fe','300':'#f0abfc','400':'#e879f9','500':'#d946ef','600':'#c026d3','700':'#a21caf','800':'#86198f','900':'#701a75','950':'#4a044e' },
-  pink: { '50':'#fdf2f8','100':'#fce7f3','200':'#fbcfe8','300':'#f9a8d4','400':'#f472b6','500':'#ec4899','600':'#db2777','700':'#be185d','800':'#9d174d','900':'#831843','950':'#500724' },
-  rose: { '50':'#fff1f2','100':'#ffe4e6','200':'#fecdd3','300':'#fda4af','400':'#fb7185','500':'#f43f5e','600':'#e11d48','700':'#be123c','800':'#9f1239','900':'#881337','950':'#4c0519' },
-};
-
-const SINGLE_COLORS: Record<string, string> = {
-  black: '#000000', white: '#ffffff', transparent: 'transparent', current: 'currentColor',
-};
-
-const COLOR_UTILITY_PREFIXES = [
-  'bg', 'text', 'border', 'ring', 'outline', 'shadow', 'divide',
-  'from', 'via', 'to', 'caret', 'accent', 'fill', 'stroke', 'decoration',
-];
-
-const NAMED_COLOR_PATTERN = new RegExp(
-  `^(${COLOR_UTILITY_PREFIXES.join('|')})-(${Object.keys(TAILWIND_COLORS).join('|')})-(50|100|200|300|400|500|600|700|800|900|950)(?:/(\\d+|\\[.+?\\]))?$`
-);
-
-const SINGLE_COLOR_PATTERN = new RegExp(
-  `^(${COLOR_UTILITY_PREFIXES.join('|')})-(${Object.keys(SINGLE_COLORS).join('|')})(?:/(\\d+|\\[.+?\\]))?$`
-);
-
-function resolveNamedColors(classes: string[]): string[] {
-  return classes.map(cls => {
-    const [variantPrefix, base] = splitVariantPrefix(cls);
-
-    const namedMatch = base.match(NAMED_COLOR_PATTERN);
-    if (namedMatch) {
-      const [, prefix, color, shade, opacity] = namedMatch;
-      const hex = TAILWIND_COLORS[color]?.[shade];
-      if (hex) {
-        return `${variantPrefix}${prefix}-[${hex}]${opacity ? `/${opacity}` : ''}`;
-      }
-    }
-
-    const singleMatch = base.match(SINGLE_COLOR_PATTERN);
-    if (singleMatch) {
-      const [, prefix, color, opacity] = singleMatch;
-      const value = SINGLE_COLORS[color];
-      if (value) {
-        if (value === 'transparent' || value === 'currentColor') {
-          return cls;
-        }
-        return `${variantPrefix}${prefix}-[${value}]${opacity ? `/${opacity}` : ''}`;
-      }
-    }
-
-    return cls;
-  });
-}
 
 // ─── Inline Style → Tailwind Classes ───
 
@@ -437,6 +237,13 @@ const ALIGN_SELF_MAP: Record<string, string> = {
   auto: 'self-auto', 'flex-start': 'self-start', start: 'self-start',
   'flex-end': 'self-end', end: 'self-end',
   center: 'self-center', stretch: 'self-stretch', baseline: 'self-baseline',
+};
+const ALIGN_CONTENT_MAP: Record<string, string> = {
+  'flex-start': 'content-start', start: 'content-start',
+  'flex-end': 'content-end', end: 'content-end',
+  center: 'content-center', 'space-between': 'content-between',
+  'space-around': 'content-around', 'space-evenly': 'content-evenly',
+  stretch: 'content-stretch',
 };
 const TEXT_ALIGN_MAP: Record<string, string> = {
   left: 'text-left', center: 'text-center', right: 'text-right', justify: 'text-justify',
@@ -493,21 +300,22 @@ function styleToClasses(style: string): string[] {
             prop === 'justify-content' ? JUSTIFY_MAP[val] :
               prop === 'align-items' ? ALIGN_ITEMS_MAP[val] :
                 prop === 'align-self' ? ALIGN_SELF_MAP[val] :
-                  prop === 'text-align' ? TEXT_ALIGN_MAP[val] :
-                    prop === 'text-decoration' || prop === 'text-decoration-line' ? TEXT_DECO_MAP[val] :
-                      prop === 'text-transform' ? TEXT_TRANSFORM_MAP[val] :
-                        prop === 'white-space' ? WHITESPACE_MAP[val] :
-                          prop === 'position' ? POSITION_MAP[val] :
-                            prop === 'overflow' ? OVERFLOW_MAP[val] :
-                              prop === 'cursor' ? CURSOR_MAP[val] :
-                                prop === 'object-fit' ? OBJECT_FIT_MAP[val] :
-                                  prop === 'font-style' && val === 'italic' ? 'italic' :
-                                    prop === 'font-style' && val === 'normal' ? 'not-italic' :
-                                      prop === 'pointer-events' && val === 'none' ? 'pointer-events-none' :
-                                        prop === 'pointer-events' && val === 'auto' ? 'pointer-events-auto' :
-                                          prop === 'word-break' && val === 'break-all' ? 'break-all' :
-                                            prop === 'overflow-wrap' && val === 'break-word' ? 'break-words' :
-                                              null;
+                  prop === 'align-content' ? ALIGN_CONTENT_MAP[val] :
+                    prop === 'text-align' ? TEXT_ALIGN_MAP[val] :
+                      prop === 'text-decoration' || prop === 'text-decoration-line' ? TEXT_DECO_MAP[val] :
+                        prop === 'text-transform' ? TEXT_TRANSFORM_MAP[val] :
+                          prop === 'white-space' ? WHITESPACE_MAP[val] :
+                            prop === 'position' ? POSITION_MAP[val] :
+                              prop === 'overflow' ? OVERFLOW_MAP[val] :
+                                prop === 'cursor' ? CURSOR_MAP[val] :
+                                  prop === 'object-fit' ? OBJECT_FIT_MAP[val] :
+                                    prop === 'font-style' && val === 'italic' ? 'italic' :
+                                      prop === 'font-style' && val === 'normal' ? 'not-italic' :
+                                        prop === 'pointer-events' && val === 'none' ? 'pointer-events-none' :
+                                          prop === 'pointer-events' && val === 'auto' ? 'pointer-events-auto' :
+                                            prop === 'word-break' && val === 'break-all' ? 'break-all' :
+                                              prop === 'overflow-wrap' && val === 'break-word' ? 'break-words' :
+                                                null;
 
     if (mapped) { classes.push(mapped); continue; }
 
@@ -583,6 +391,21 @@ function styleToClasses(style: string): string[] {
         break;
       case 'box-shadow':
         classes.push(`shadow-[${val.replace(/\s+/g, '_')}]`);
+        break;
+      case 'background-image':
+        classes.push(`bg-[${val.replace(/\s+/g, '_')}]`);
+        break;
+      case 'flex-grow':
+        classes.push(val === '0' ? 'grow-0' : 'grow');
+        break;
+      case 'flex-shrink':
+        classes.push(val === '0' ? 'shrink-0' : 'shrink');
+        break;
+      case 'flex-basis':
+        classes.push(val === 'auto' ? 'basis-auto' : `basis-[${val}]`);
+        break;
+      case 'order':
+        classes.push(`order-[${val}]`);
         break;
     }
   }
@@ -703,13 +526,13 @@ function makeRichTextVariable(textOrDoc: string | object) {
   };
 }
 
-function makeTextLayer(text: string): Layer {
+function makeTextLayer(textOrDoc: string | object): Layer {
   return {
     id: generateId('lyr'),
     name: 'text',
     classes: '',
     restrictions: { editText: true },
-    variables: { text: makeRichTextVariable(text) },
+    variables: { text: makeRichTextVariable(textOrDoc) },
   };
 }
 
@@ -745,10 +568,25 @@ function resolveImportClasses(el: Element): string {
   return resolved.join(' ');
 }
 
+function sanitizeSvg(el: Element): void {
+  el.querySelectorAll('script').forEach(s => s.remove());
+  const walk = (node: Element) => {
+    for (const attr of Array.from(node.attributes)) {
+      if (attr.name.toLowerCase().startsWith('on')) {
+        node.removeAttribute(attr.name);
+      }
+    }
+    for (let i = 0; i < node.children.length; i++) {
+      walk(node.children[i]);
+    }
+  };
+  walk(el);
+}
+
 function elementToLayer(el: Element): Layer | null {
   const tag = el.tagName.toLowerCase();
 
-  if (tag === 'script' || tag === 'style' || tag === 'link' || tag === 'meta') {
+  if (tag === 'script' || tag === 'style' || tag === 'link' || tag === 'meta' || tag === 'br') {
     return null;
   }
 
@@ -828,6 +666,38 @@ function elementToLayer(el: Element): Layer | null {
     return layer;
   }
 
+  if (tag === 'textarea') {
+    const placeholder = el.getAttribute('placeholder');
+    const name = el.getAttribute('name');
+    const rows = el.getAttribute('rows');
+    layer.attributes = {
+      ...layer.attributes,
+      ...(placeholder ? { placeholder } : {}),
+      ...(name ? { name } : {}),
+      ...(rows ? { rows } : {}),
+    };
+    return layer;
+  }
+
+  if (tag === 'select') {
+    const name = el.getAttribute('name');
+    layer.attributes = {
+      ...layer.attributes,
+      ...(name ? { name } : {}),
+    };
+    return layer;
+  }
+
+  if (tag === 'form') {
+    const action = el.getAttribute('action');
+    const method = el.getAttribute('method');
+    layer.attributes = {
+      ...layer.attributes,
+      ...(action ? { action } : {}),
+      ...(method ? { method } : {}),
+    };
+  }
+
   if (tag === 'iframe') {
     const src = el.getAttribute('src');
     if (src) {
@@ -862,6 +732,7 @@ function elementToLayer(el: Element): Layer | null {
   }
 
   if (tag === 'svg') {
+    sanitizeSvg(el);
     const svgString = el.outerHTML;
     layer.variables = {
       ...layer.variables,
@@ -892,14 +763,7 @@ function elementToLayer(el: Element): Layer | null {
   if (!isTextLayer && isTextOnlyElement(el) && el.textContent?.trim()) {
     const doc = buildRichTextDoc(el);
     if (CONTAINER_NAMES.has(layerName)) {
-      const textLayer: Layer = {
-        id: generateId('lyr'),
-        name: 'text',
-        classes: '',
-        restrictions: { editText: true },
-        variables: { text: makeRichTextVariable(doc) },
-      };
-      layer.children = [textLayer];
+      layer.children = [makeTextLayer(doc)];
     } else {
       layer.variables = {
         ...layer.variables,
@@ -949,37 +813,34 @@ function elementToLayer(el: Element): Layer | null {
 export function htmlToLayers(html: string): Layer[] {
   if (typeof window === 'undefined') return [];
 
-  const parser = new DOMParser();
-  const doc = parser.parseFromString(html, 'text/html');
+  try {
+    const parser = new DOMParser();
+    const doc = parser.parseFromString(html, 'text/html');
 
-  const layers: Layer[] = [];
-  const body = doc.body;
+    const layers: Layer[] = [];
+    const body = doc.body;
 
-  for (let i = 0; i < body.childNodes.length; i++) {
-    const node = body.childNodes[i];
-    if (node.nodeType === Node.ELEMENT_NODE) {
-      const layer = elementToLayer(node as Element);
-      if (layer) layers.push(layer);
-    } else if (node.nodeType === Node.TEXT_NODE) {
-      const text = (node.textContent || '').trim();
-      if (text) {
-        layers.push(makeTextLayer(text));
+    for (let i = 0; i < body.childNodes.length; i++) {
+      const node = body.childNodes[i];
+      if (node.nodeType === Node.ELEMENT_NODE) {
+        const layer = elementToLayer(node as Element);
+        if (layer) layers.push(layer);
+      } else if (node.nodeType === Node.TEXT_NODE) {
+        const text = (node.textContent || '').trim();
+        if (text) {
+          layers.push(makeTextLayer(text));
+        }
       }
     }
-  }
 
-  return layers;
+    return layers;
+  } catch (err) {
+    console.warn('htmlToLayers: failed to parse HTML', err);
+    return [];
+  }
 }
 
 // ─── Export: Layers → HTML ───
-
-function escapeHtml(str: string): string {
-  return str
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;');
-}
 
 const MARK_TO_HTML_TAG: Record<string, string> = {
   bold: 'strong', italic: 'em', underline: 'u', strike: 's',
@@ -1095,6 +956,21 @@ function layerToHtmlString(layer: Layer, indent: number): string {
     if (layer.attributes?.type) attrs.push(`type="${escapeHtml(layer.attributes.type)}"`);
     if (layer.attributes?.placeholder) attrs.push(`placeholder="${escapeHtml(layer.attributes.placeholder)}"`);
     if (layer.attributes?.name) attrs.push(`name="${escapeHtml(layer.attributes.name)}"`);
+  }
+
+  if (layer.name === 'textarea') {
+    if (layer.attributes?.placeholder) attrs.push(`placeholder="${escapeHtml(layer.attributes.placeholder)}"`);
+    if (layer.attributes?.name) attrs.push(`name="${escapeHtml(layer.attributes.name)}"`);
+    if (layer.attributes?.rows) attrs.push(`rows="${escapeHtml(String(layer.attributes.rows))}"`);
+  }
+
+  if (layer.name === 'select') {
+    if (layer.attributes?.name) attrs.push(`name="${escapeHtml(layer.attributes.name)}"`);
+  }
+
+  if (layer.name === 'form') {
+    if (layer.attributes?.action) attrs.push(`action="${escapeHtml(layer.attributes.action)}"`);
+    if (layer.attributes?.method) attrs.push(`method="${escapeHtml(layer.attributes.method)}"`);
   }
 
   if (layer.name === 'iframe') {
